requirements.txt
aiohttp==3.9.5
aiokafka~=0.12.0
asyncpg==0.29.0
alembic==1.13.2
fastapi==0.111.0
greenlet==3.0.3
gunicorn==22.0.0
httpx~=0.28.1
pydantic==2.7.1
pydantic-core==2.18.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
uvicorn==0.29.0
SQLAlchemy~=2.0.31
